MQTT-LogApi/requirements.txt
async-timeout==4.0.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
eventlet==0.33.3
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SSE==1.0.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
paho-mqtt==1.6.1
python-dotenv==1.0.0
redis==4.5.5
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
urllib3==2.0.2
Werkzeug==2.3.4
